requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docstring-to-markdown==0.12
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.7
flake8==6.0.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
fqdn==1.5.1
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
lazy-object-proxy==1.9.0
lxml==4.9.2
macholib==1.16.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
multidict==6.0.4
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.2.3
pyflakes==3.0.1
Pygments==2.14.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pylint==2.17.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
pytoolconfig==1.2.5
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rope==1.7.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
snscrape==0.6.2.20230320
soupsieve==2.4
stack-data==0.6.2
telnetlib3==2.0.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomlkit==0.11.6
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
ujson==5.7.0
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
whatthepatch==1.0.4
widgetsnbextension==4.0.7
wrapt==1.15.0
yarl==1.8.2
redis==4.5.4
rq==1.13.0
sqlalchemy==2.0.8
psycopg2-binary==2.9.6
Flask-JWT-Extended==4.4.4
PyJWT==2.6.0
Flask-SQLAlchemy==3.0.3
Flask-Redis==0.4.0
python-http-client==3.3.7
sendgrid==6.10.0
starkbank-ecdsa==2.2.0
flask_migrate==4.0.4
alembic==1.10.2
Flask-Login==0.6.2
marshmallow-sqlalchemy==0.29.0
algoliasearch==3.0.0
boto3==1.26.110
botocore==1.29.110
jmespath==1.0.1
s3transfer==0.6.0
opencc-python-reimplemented==0.1.7
pillow==9.5.0
babel==2.12.1
aniso8601==9.0.1
tenacity==8.2.2
crontab==1.0.1
freezegun==1.2.2
rq-scheduler==0.13.0
dataclasses-json==0.5.7
langchain==0.0.157
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
numexpr==2.8.4
openapi-schema-pydantic==1.2.4
pydantic==1.10.7
typing-inspect==0.8.0